Should fail the test cases split to a browser that is disconnected
Each time the browser unregister and re-register his shard id become different which causes round-robin to not give tests to this browser instance.

The fix is to find available shard index (in our case from 0 - 3) and assign it to the new browser instance.
For some reason the shard index first time is number and there are no issues in round-robin function but the second time (re-connect) it is string which reflects to no tests for this browser, because his index doesn't "answering" the needed condiditon. For example: 15 % 3 === '0' -> False. The fix is to do Number(a) === Number(b).